pyproject.toml
[project]
name = "consolidate-bear-markdown"
version = "0.1.0"
description = "Process Bear.app markdown files and attachments with GPT-4o vision analysis"
authors = [{ name = "Chad Walters", email = "chad.walters@gmail.com" }]
dependencies = [
  "openai",               # For GPT-4o vision model
  "wand",                 # For SVG to PNG conversion
  "pillow",               # For image processing
  "toml",                 # For configuration files
  "tqdm",                 # For progress bars
  "pandas",               # For data processing
  "tabulate",             # For markdown tables
  "beautifulsoup4",       # For HTML processing
  "types-beautifulsoup4", # Type stubs for beautifulsoup4
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
  "pytest",     # Core testing framework
  "pytest-cov", # Code coverage reporting
  "mypy",       # Type checking
  "black",      # Code formatting
  "ruff",       # Fast linting
  "pre-commit", # Pre-commit hooks
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
tmp_path_retention_count = 0
tmp_path_retention_policy = "failed"
filterwarnings = [
  "default",                                                      # Show all warnings by default
  "ignore::DeprecationWarning:_frozen_importlib_external.*:",     # Ignore importlib warnings
  "ignore::DeprecationWarning:importlib._bootstrap.*:",           # Ignore bootstrap warnings
  "ignore::DeprecationWarning:.*:0",                              # Ignore all other deprecation warnings
  "ignore:builtin type SwigPyPacked has no __module__ attribute",
  "ignore:builtin type SwigPyObject has no __module__ attribute",
  "ignore:builtin type swigvarlink has no __module__ attribute",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors
  "F", # pyflakes
  "B", # flake8-bugbear
  "I", # isort
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F841"] # Ignore unused variables in tests
